fix & clean up cavern creature classes

+++ src/main/java/quek/undergarden/entity/cavern/SploogieEntity.java
@@ -1,29 +1,26 @@
 package quek.undergarden.entity.cavern;
 
-import net.minecraft.entity.*;
-import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
-import net.minecraft.world.entity.ai.attributes.Attributes;
-import net.minecraft.entity.ai.goal.*;
-import net.minecraft.world.entity.monster.Monster;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.util.Mth;
-import net.minecraft.world.level.Level;
-import quek.undergarden.entity.projectile.SlingshotAmmoEntity;
-import quek.undergarden.registry.UGSoundEvents;
-
+import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EntityDimensions;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Pose;
+import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.ai.goal.FloatGoal;
 import net.minecraft.world.entity.ai.goal.PanicGoal;
 import net.minecraft.world.entity.ai.goal.RangedAttackGoal;
 import net.minecraft.world.entity.ai.goal.WaterAvoidingRandomStrollGoal;
 import net.minecraft.world.entity.ai.goal.target.HurtByTargetGoal;
 import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
+import net.minecraft.world.entity.monster.Monster;
 import net.minecraft.world.entity.monster.RangedAttackMob;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import quek.undergarden.entity.projectile.SlingshotAmmoEntity;
+import quek.undergarden.registry.UGSoundEvents;
 
 public class SploogieEntity extends AbstractCavernCreatureEntity implements RangedAttackMob {
 
@@ -68,7 +65,7 @@ public class SploogieEntity extends AbstractCavernCreatureEntity implements Rang
         double xDistance = target.getX() - this.getX();
         double yDistance = target.getY(0.3333333333333333D) - pebble.getY();
         double zDistance = target.getZ() - this.getZ();
-        double yMath = Mth.sqrt((xDistance * xDistance) + (zDistance * zDistance));
+        double yMath = Mth.sqrt((float) ((xDistance * xDistance) + (zDistance * zDistance)));
         pebble.shoot(xDistance, yDistance + yMath * 0.1D, zDistance, 1.6F, 1.0F);
         this.playSound(UGSoundEvents.SPLOOGIE_SPIT.get(), 1.0F, this.getVoicePitch());
         this.level.addFreshEntity(pebble);
